@simonmichael 2018-05-24
It is more verbose, but clarity (and performance!) is good. Thanks.